+++ ros2controlcli/ros2controlcli/api/__init__.py
@@ -14,10 +14,7 @@
 
 
 from controller_manager_msgs.srv import \
-    ConfigureController, ConfigureStartController, \
-    ListControllers, ListControllerTypes, ListHardwareInterfaces, \
-    LoadConfigureController, LoadController, LoadStartController, \
-    ReloadControllerLibraries, SwitchController, UnloadController
+    ConfigureStartController, LoadConfigureController, LoadStartController
 
 import rclpy
 
@@ -57,47 +54,6 @@ def service_caller(service_name, service_type, request):
         rclpy.shutdown()
 
 
-def list_controllers(controller_manager_name):
-    request = ListControllers.Request()
-    return service_caller('{}/list_controllers'.format(controller_manager_name),
-                          ListControllers, request)
-
-
-def list_controller_types(controller_manager_name):
-    request = ListControllerTypes.Request()
-    return service_caller(
-        '{}/list_controller_types'.format(controller_manager_name), ListControllerTypes, request)
-
-
-def list_hardware_interfaces(controller_manager_name):
-    request = ListHardwareInterfaces.Request()
-    return service_caller('{}/list_hardware_interfaces'.format(controller_manager_name),
-                          ListHardwareInterfaces, request)
-
-
-def reload_controller_libraries(controller_manager_name, force_kill):
-    request = ReloadControllerLibraries.Request()
-    request.force_kill = force_kill
-    return service_caller(
-        '{}/reload_controller_libraries'.format(controller_manager_name),
-        ReloadControllerLibraries,
-        request)
-
-
-def load_controller(controller_manager_name, controller_name):
-    request = LoadController.Request()
-    request.name = controller_name
-    return service_caller('{}/load_controller'.format(controller_manager_name),
-                          LoadController, request)
-
-
-def configure_controller(controller_manager_name, controller_name):
-    request = ConfigureController.Request()
-    request.name = controller_name
-    return service_caller('{}/configure_controller'.format(controller_manager_name),
-                          ConfigureController, request)
-
-
 def load_configure_controller(controller_manager_name, controller_name):
     request = LoadConfigureController.Request()
     request.name = controller_name
@@ -117,28 +73,6 @@ def configure_start_controller(controller_manager_name, controller_name):
     request.name = controller_name
     return service_caller('{}/configure_and_start_controller'.format(controller_manager_name),
                           ConfigureStartController, request)
-
-
-def switch_controllers(controller_manager_name, stop_controllers,
-                       start_controllers, strict, start_asap, timeout):
-    request = SwitchController.Request()
-    request.start_controllers = start_controllers
-    request.stop_controllers = stop_controllers
-    if strict:
-        request.strictness = SwitchController.Request.STRICT
-    else:
-        request.strictness = SwitchController.Request.BEST_EFFORT
-    request.start_asap = start_asap
-    request.timeout = rclpy.duration.Duration(seconds=timeout).to_msg()
-    return service_caller('{}/switch_controller'.format(controller_manager_name),
-                          SwitchController, request)
-
-
-def unload_controller(controller_manager_name, controller_name):
-    request = UnloadController.Request()
-    request.name = controller_name
-    return service_caller('{}/unload_controller'.format(controller_manager_name),
-                          UnloadController, request)
 
 
 class ControllerNameCompleter:
